Complete basic dns server
This commit consolidates the dns server code with the scheduled pinger
into one complete mini-dns server. It is a mulithreaded udp server which
accepts dns queries seeking the domain
'snl-columbia-university.github.io', and returns a dns query with an
answer to an IP address with the least latency (per the most recent ping
to the available servers). It logs the results of the ping to a log
file, along with all recieved and outgoing dns queries. Since it's
multithreaded, it locks the object containing the latency information,
as well as the output log file.